FP-0009 users-module
