+++ src/app/users/users-routing.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { UserEditComponent } from './user-edit/user-edit.component';
+import { UsersComponent } from './users.component';
+
+const routes: Routes = [
+  {
+    path: 'users',
+    component: UsersComponent,
+  },
+  { path: 'users/:id/edit', component: UserEditComponent },
+  { path: 'users/new', component: UserEditComponent },
+];
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class UsersRoutingModule {}
