+++ src/app/users/users.module.ts
@@ -1,0 +1,30 @@
+import { NgModule } from '@angular/core';
+import { UserDeleteConfirm, UsersComponent } from './users.component';
+import { UserEditComponent } from './user-edit/user-edit.component';
+import { RouterModule } from '@angular/router';
+import { UsersRoutingModule } from './users-routing.module';
+import { FormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { MatListModule } from '@angular/material/list';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatTableModule } from '@angular/material/table';
+import { MatDialogModule } from '@angular/material/dialog';
+
+@NgModule({
+  declarations: [UsersComponent, UserEditComponent, UserDeleteConfirm],
+  imports: [
+    FormsModule,
+    CommonModule,
+    RouterModule,
+    UsersRoutingModule,
+    MatListModule,
+    MatToolbarModule,
+    MatIconModule,
+    MatButtonModule,
+    MatTableModule,
+    MatDialogModule,
+  ],
+})
+export class UsersModule {}
